Add files via upload
This project contains the Ola Bike dataset, through overview and cleaning of the dataset, visualization using plotly, encoding, scaling and the use of different machine learning Regression models to train and test this data for proper and best accuracy for Ola Bike prediction request. deployment using Streamlit (VS code) to create a prediction app. And creation of a Dashboard for visualization and interaction with the cleaned data using Power BI.